We expect iov lengths to match up with a size_t, not ssize_t.
git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@568881 13f79535-47bb-0310-9956-ffa450edef68

+++ include/apr_want.h
@@ -92,7 +92,7 @@
 struct iovec
 {
     char *iov_base;
-    int iov_len;
+    size_t iov_len;
 };
 
 #endif
